requirements.txt
google-api-python-client==2.42.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
python-dotenv==0.20.0